Some refactors and worker code improvement
With this new version the worker job is improved in legibility and performance. Some problems occur again but the result is closer to the final implementation then the past. At this point the time required in order to receive the response is about 10x then the to obtain the same result without pass through the web switch (we talk about 10ms for direct connection and 100ms for webswitched one).