Resolves problems with some request and it improves print on the terminal
It was fixed a problem with http_request header parser that did not parse the last line (sometimes the line was a Content-lenght line and so a Bad Address was thrown when the worker read). It is also changed all useful prints with log message and it was introduced a function that helps with unused parameter.

In order to use this function include macro.h and when a variable is unused "wrap" it with "UNUSED(unused_variable)".